+++ instrumentation/clickhouse/clickhouse-client-0.5/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/clickhouse/v0_5/ClickHouseClientV1InstrumentationModule.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.clickhouse.v0_5;
+
+import static java.util.Collections.singletonList;
+
+import com.google.auto.service.AutoService;
+import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
+import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
+import io.opentelemetry.javaagent.extension.instrumentation.internal.ExperimentalInstrumentationModule;
+import java.util.List;
+
+@AutoService(InstrumentationModule.class)
+public class ClickHouseClientV1InstrumentationModule extends InstrumentationModule
+    implements ExperimentalInstrumentationModule {
+
+  public ClickHouseClientV1InstrumentationModule() {
+    super("clickhouse-client", "clickhouse-client-0.5", "clickhouse");
+  }
+
+  @Override
+  public boolean isHelperClass(String className) {
+    return "com.clickhouse.client.ClickHouseRequestAccess".equals(className);
+  }
+
+  @Override
+  public List<String> injectedClassNames() {
+    return singletonList("com.clickhouse.client.ClickHouseRequestAccess");
+  }
+
+  @Override
+  public List<TypeInstrumentation> typeInstrumentations() {
+    return singletonList(new ClickHouseClientV1Instrumentation());
+  }
+}
