+++ instrumentation/clickhouse/clickhouse-client-0.6/javaagent/build.gradle.kts
@@ -1,0 +1,41 @@
+plugins {
+  id("otel.javaagent-instrumentation")
+}
+
+muzzle {
+  pass {
+    group.set("com.clickhouse")
+    module.set("client-v2")
+    versions.set("[0.6.4,)")
+    assertInverse.set(true)
+  }
+}
+
+dependencies {
+  implementation(project(":instrumentation:clickhouse:clickhouse-client-common:javaagent"))
+  compileOnly("com.clickhouse:client-v2:0.6.4")
+
+  testImplementation("com.google.guava:guava")
+  testImplementation(project(":instrumentation:clickhouse:testing"))
+  testLibrary("com.clickhouse:client-v2:0.6.4")
+}
+
+val collectMetadata = findProperty("collectMetadata")?.toString() ?: "false"
+
+tasks {
+  test {
+    usesService(gradle.sharedServices.registrations["testcontainersBuildService"].service)
+    systemProperty("collectMetadata", collectMetadata)
+  }
+
+  val testStableSemconv by registering(Test::class) {
+    jvmArgs("-Dotel.semconv-stability.opt-in=database")
+
+    systemProperty("metaDataConfig", "otel.semconv-stability.opt-in=database")
+    systemProperty("collectMetadata", collectMetadata)
+  }
+
+  check {
+    dependsOn(testStableSemconv)
+  }
+}
