+++ instrumentation/clickhouse/clickhouse-client-0.6/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/clickhouse/v0_6/ClickHouseClientV2Instrumentation.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.clickhouse.v0_6;
+
+import static net.bytebuddy.matcher.ElementMatchers.isMethod;
+import static net.bytebuddy.matcher.ElementMatchers.isSubTypeOf;
+import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
+
+import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
+import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
+import java.util.Map;
+import net.bytebuddy.description.type.TypeDescription;
+import net.bytebuddy.matcher.ElementMatcher;
+
+public class ClickHouseClientV2Instrumentation implements TypeInstrumentation {
+  @Override
+  public ElementMatcher<TypeDescription> typeMatcher() {
+    return named("com.clickhouse.client.api.Client");
+  }
+
+  @Override
+  public void transform(TypeTransformer transformer) {
+    transformer.applyAdviceToMethod(
+        isMethod()
+            .and(named("query"))
+            .and(takesArgument(0, String.class))
+            .and(takesArgument(1, isSubTypeOf(Map.class)))
+            .and(takesArgument(2, named("com.clickhouse.client.api.query.QuerySettings"))),
+        "io.opentelemetry.javaagent.instrumentation.clickhouse.v0_6.ClickHouseClientV2QueryAdvice");
+  }
+}
