+++ instrumentation/clickhouse/clickhouse-client-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/clickhouse/common/ClickHouseInstrumenterFactory.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.clickhouse.common;
+
+import io.opentelemetry.api.GlobalOpenTelemetry;
+import io.opentelemetry.instrumentation.api.incubator.semconv.db.DbClientAttributesExtractor;
+import io.opentelemetry.instrumentation.api.incubator.semconv.db.DbClientMetrics;
+import io.opentelemetry.instrumentation.api.incubator.semconv.db.DbClientSpanNameExtractor;
+import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
+import io.opentelemetry.instrumentation.api.instrumenter.SpanKindExtractor;
+import io.opentelemetry.instrumentation.api.semconv.network.ServerAttributesExtractor;
+
+public final class ClickHouseInstrumenterFactory {
+
+  public static Instrumenter<ClickHouseDbRequest, Void> createInstrumenter(
+      String instrumenterName) {
+    ClickHouseAttributesGetter dbAttributesGetter = new ClickHouseAttributesGetter();
+
+    return Instrumenter.<ClickHouseDbRequest, Void>builder(
+            GlobalOpenTelemetry.get(),
+            instrumenterName,
+            DbClientSpanNameExtractor.create(dbAttributesGetter))
+        .addAttributesExtractor(DbClientAttributesExtractor.create(dbAttributesGetter))
+        .addAttributesExtractor(
+            ServerAttributesExtractor.create(new ClickHouseNetworkAttributesGetter()))
+        .addOperationMetrics(DbClientMetrics.get())
+        .buildInstrumenter(SpanKindExtractor.alwaysClient());
+  }
+
+  private ClickHouseInstrumenterFactory() {}
+}
