+++ settings.gradle.kts
@@ -190,7 +190,10 @@ include(":instrumentation:cassandra:cassandra-4.4:library")
 include(":instrumentation:cassandra:cassandra-4.4:testing")
 include(":instrumentation:cassandra:cassandra-4-common:testing")
 include(":instrumentation:cdi-testing")
-include(":instrumentation:clickhouse-client-0.5:javaagent")
+include(":instrumentation:clickhouse:clickhouse-client-common:javaagent")
+include(":instrumentation:clickhouse:clickhouse-client-0.6:javaagent")
+include(":instrumentation:clickhouse:clickhouse-client-0.5:javaagent")
+include(":instrumentation:clickhouse:testing")
 include(":instrumentation:couchbase:couchbase-2.0:javaagent")
 include(":instrumentation:couchbase:couchbase-2.6:javaagent")
 include(":instrumentation:couchbase:couchbase-2-common:javaagent")
